refactor(zoominfo): modularize connector, drop unused test, fix lint
Address Fivetran engineer review feedback on PR #23:

- Split the 2610-line connector.py into focused modules, mirroring the
  callminer/ multi-module precedent (connector.py is still the entry
  point Fivetran loads):
    constants.py   — endpoint paths, tunables, default field lists, state keys
    client.py      — OAuth token cache, retry transport, pagination
    config.py      — config parsing/validation, search filter, eligibility
    transforms.py  — pure value coercion helpers
    sync.py        — one sync_* per table + streaming enrichment helper
    connector.py   — schema() + update() + Connector object (472 lines)
- Revert the module-level constants to plain UPPER names now that they
  live in a shared constants.py imported by name (matches ehi/ precedent).
- Remove connector_test.py (not used by the connector) and its caches.
- Pass the repo code-quality checks locally: black --line-length 99 and
  flake8 both clean (rewrapped the long comment/docstring lines that
  tripped B950).
- Expand the root README.md zoominfo entry and document the module
  layout in zoominfo/README.md.

No behavior change — pure code movement plus line-wrapping and renames.
